fix(core): stabilize build — remove warnings, fix deprecated API, deduplicate routes, stable Hasql pool
- Fix Hasql statement usage in DB.Document.RegisterType and DB.Document.Audit
- Remove unused functions in Surypus.Reports.Templates
- Remove unused imports in Surypus.Types, Surypus.Z3, Core.Payroll.Calculation, Core.Warehouse
- Fix type defaulting in Surypus.Types
- Fix variable shadowing in Core.Production.Types, Core.QCert, Core.SmartReceipt
- Fix toJSON derivation in CrystalToPdfSlave
- Temporarily disable missing converters in CrystalToJasper

+++ src/DB/User.hs
@@ -42,8 +42,8 @@ userRow =
     <*> column (D.nonNullable D.text)
     <*> column (D.nonNullable D.text)
     <*> column (D.nonNullable D.text)
-    <*> column (D.nonNullable D.int4)
-    <*> column (D.nonNullable D.int4)
+    <*> column (D.nonNullable D.int8)
+    <*> column (D.nonNullable D.int8)
 
 rolesFromFlags :: Int64 -> Text
 rolesFromFlags flags
